ai/requirements.txt
torch==2.2.2
torchvision==0.17.2
onnx==1.16.0
onnxruntime==1.17.3
pillow==10.3.0
numpy==1.26.4
scikit-learn==1.4.2
